requirements.txt
azure-identity
azure-storage-blob
boto3
google-cloud-storage
oauth2client
